backend/requirements.txt
fastapi==0.115.0
mangum==0.19.0
uvicorn==0.30.0
boto3==1.35.0
Pillow==10.4.0
python-jose[cryptography]==3.3.0
requests==2.32.0
python-multipart==0.0.9
razorpay
aiofiles==23.2.1
